refactor(server): extract shutdown into src/shutdown.rs
Moves shutdown_signal, select_shutdown_trigger, arm_force_exit,
force_exit, and SignalListener (plus their unit tests) out of
src/main.rs into a new src/shutdown.rs, per src/AGENTS.md's Project
Structure and 500-line guidance. Purely mechanical: serve() now calls
shutdown::shutdown_signal() instead of a same-file function, no logic
changed.
